backend/API_Avaliauto/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.1.0
django-cors-headers
